integrate with raw bson visitor
