+++ src/raw/bson.rs
@@ -24,7 +24,7 @@ use crate::{
 };
 
 use super::{
-    serde::{OwnedOrBorrowedRawBson, OwnedOrBorrowedRawBsonVisitor},
+    serde::{bson_visitor::OwnedOrBorrowedRawBsonVisitor, OwnedOrBorrowedRawBson},
     Error,
     Result,
 };
